@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import Box from '@material-ui/core/Box';
-import Typography from '@material-ui/core/Typography';
+import Box from '@mui/material/Box';
+import Typography from '@mui/material/Typography';
 
 const FormHeader = ({ children, disablePadding, ...rest }) => (
   <Box color='text.secondary' mt={disablePadding ? 0 : 2} mb={0.5} {...rest}>
