@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Box from '@material-ui/core/Box';
-import Divider from '@material-ui/core/Divider';
+import Box from '@mui/material/Box';
+import Divider from '@mui/material/Divider';
 
 /**
  * Divider that fits nicely in a mini list.
