@@ -1,4 +1,4 @@
 /**
  * Helper function that returns whether the given Material UI theme is a dark theme.
  */
-export const isThemeDark = (theme) => theme.palette.type === 'dark';
+export const isThemeDark = (theme) => theme.palette.mode === 'dark';
